refactor: remove unnecessary IrStatementLists.kt
